test: move lib tests to test_lib.rs
